+++ backend-rust/CHANGELOG.md
@@ -11,6 +11,10 @@ Database schema version: 4
 - Database migration to add the lottery power of each baker pool during the last payday period.
 - Add Query `Query::Baker::state::pool::lotteryPower` which returns the `lotteryPower` of the baker pool during the last payday period.
 
+### Changed
+
+- Query the `get_module_source` on the `LastFinal` block to improve performance.
+
 ## [0.1.24] - 2025-02-12
 
 ### Fixed